requirements.txt
streamlit>=1.32
pandas>=2.0
python-dateutil>=2.8
openpyxl>=3.1
reportlab>=4.0
